requirements.txt
pandas
openpyxl
discord.py